Fix form-related circular imports
